SIG-3904 split models module
